Add onRequest route hook, refactor tests and framework for route hooks
